+++ src/main/java/roomescape/controller/AuthController.java
@@ -1,0 +1,57 @@
+package roomescape.controller;
+
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.service.AuthService;
+import roomescape.service.dto.LoginMember;
+import roomescape.service.dto.request.LoginRequest;
+import roomescape.service.dto.response.MemberResponse;
+
+@RestController
+public class AuthController {
+
+    public static final String TOKEN_NAME = "token";
+
+    private final AuthService authService;
+
+    public AuthController(AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<Void> login(@Valid @RequestBody LoginRequest loginRequest, HttpServletResponse response) {
+        String token = authService.login(loginRequest);
+
+        Cookie cookie = new Cookie(TOKEN_NAME, token);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        response.addCookie(cookie);
+
+        return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/login/check")
+    public ResponseEntity<MemberResponse> findMember(LoginMember loginMember) {
+        if (loginMember == null) {
+            return ResponseEntity.ok().build();
+        }
+
+        MemberResponse memberResponse = new MemberResponse(loginMember.getId(), loginMember.getName());
+        return ResponseEntity.ok(memberResponse);
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(HttpServletResponse response) {
+        Cookie cookie = new Cookie(TOKEN_NAME, null);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        response.addCookie(cookie);
+        return ResponseEntity.ok().build();
+    }
+}
