+++ src/main/java/roomescape/controller/MemberController.java
@@ -1,59 +1,19 @@
 package roomescape.controller;
 
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.service.MemberService;
-import roomescape.service.dto.LoginMember;
-import roomescape.service.dto.request.LoginRequest;
 import roomescape.service.dto.response.MemberResponse;
 
 @RestController
 public class MemberController {
 
-    public static final String TOKEN_NAME = "token";
-
     private final MemberService memberService;
 
     public MemberController(MemberService memberService) {
         this.memberService = memberService;
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<Void> login(@Valid @RequestBody LoginRequest loginRequest, HttpServletResponse response) {
-        String token = memberService.login(loginRequest);
-
-        Cookie cookie = new Cookie(TOKEN_NAME, token);
-        cookie.setHttpOnly(true);
-        cookie.setPath("/");
-        response.addCookie(cookie);
-
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/login/check")
-    public ResponseEntity<MemberResponse> findMember(LoginMember loginMember) {
-        if (loginMember == null) {
-            return ResponseEntity.ok().build();
-        }
-
-        MemberResponse memberResponse = new MemberResponse(loginMember.getId(), loginMember.getName());
-        return ResponseEntity.ok(memberResponse);
-    }
-
-    @PostMapping("/logout")
-    public ResponseEntity<Void> logout(HttpServletResponse response) {
-        Cookie cookie = new Cookie(TOKEN_NAME, null);
-        cookie.setHttpOnly(true);
-        cookie.setPath("/");
-        response.addCookie(cookie);
-        return ResponseEntity.ok().build();
     }
 
     @GetMapping("/members")
